+++ src/test/java/kitchenpos/domain/ProductsTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.product;
+package kitchenpos.domain;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
